requirements.txt
celery==4.1.0
python-libnmap
requests
IPy